Remove SDK helpers
Also, iterate over schema files instead of manually importing each one.
