+++ jest.config.js
@@ -11,16 +11,14 @@ module.exports = {
   clearMocks: true,
   collectCoverage: true,
   collectCoverageFrom: [
-    "src/**/*.ts",
-    "!src/converters.ts",
-    "!src/ip.ts"
+    "src/**/*.ts"
   ],
   coverageThreshold: {
     global: {
-      statements: 97,
-      branches: 87,
+      statements: 100,
+      branches: 100,
       functions: 100,
-      lines: 97,
+      lines: 100,
     },
   },
 };
