+++ src/RelationshipClass.ts
@@ -1,0 +1,23 @@
+export enum RelationshipClass {
+  ALLOWS = 'ALLOWS',
+  ASSIGNED = 'ASSIGNED',
+  CONNECTS = 'CONNECTS',
+  CONTAINS = 'CONTAINS',
+  DEPLOYED = 'DEPLOYED',
+  EXTENDS = 'EXTENDS',
+  HAS = 'HAS',
+  IDENTIFIED = 'IDENTIFIED',
+  IMPLEMENTS = 'IMPLEMENTS',
+  IS = 'IS',
+  MANAGES = 'MANAGES',
+  MITIGATES = 'MITIGATES',
+  MONITORS = 'MONITORS',
+  OPENED = 'OPENED',
+  OWNS = 'OWNS',
+  PERFORMED = 'PERFORMED',
+  PROTECTS = 'PROTECTS',
+  PROVIDES = 'PROVIDES',
+  TRIGGERS = 'TRIGGERS',
+  TRUSTS = 'TRUSTS',
+  USES = 'USES',
+};
