+++ tsconfig.json
@@ -15,7 +15,6 @@
     "target": "ESNext",
     "module": "commonjs",
     "pretty": true,
-    "inlineSourceMap": true,
-    "resolveJsonModule": true
+    "inlineSourceMap": true
   }
 }
